LLM/requirements.txt
# requirements.txt
openai==1.3.5
python-dotenv==1.0.0
argparse==1.4.0